Move getPatientId up the chain
