+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/external/hapi/HapiMessage.java
@@ -5,6 +5,8 @@ import gov.hhs.cdc.trustedintermediary.etor.messages.Message;
 import gov.hhs.cdc.trustedintermediary.etor.messages.MessageHdDataType;
 import java.util.function.Supplier;
 import org.hl7.fhir.r4.model.Bundle;
+import org.hl7.fhir.r4.model.Identifier;
+import org.hl7.fhir.r4.model.Patient;
 
 public class HapiMessage implements Message<Bundle> {
 
@@ -62,6 +64,21 @@ public class HapiMessage implements Message<Bundle> {
                 () -> MESSAGE_HELPER.extractReceivingFacilityNamespace(innerResource),
                 () -> MESSAGE_HELPER.extractReceivingFacilityUniversalId(innerResource),
                 () -> MESSAGE_HELPER.extractReceivingFacilityUniversalIdType(innerResource));
+    }
+
+    public String getPatientId() {
+        return HapiHelper.resourcesInBundle(innerResource, Patient.class)
+                .flatMap(patient -> patient.getIdentifier().stream())
+                .filter(
+                        identifier ->
+                                identifier
+                                        .getType()
+                                        .hasCoding(
+                                                "http://terminology.hl7.org/CodeSystem/v2-0203",
+                                                "MR"))
+                .map(Identifier::getValue)
+                .findFirst()
+                .orElse("");
     }
 
     protected MessageHdDataType extractMessageHdDataType(
